requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and models
pydantic==2.5.0

# HTTP client for API calls
requests==2.32.3

# Database
sqlalchemy==2.0.23
alembic>=1.13.0

# Authentication
PyJWT>=2.8.0

# Google Cloud
google-cloud-secret-manager>=2.16.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# OpenAI API
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0
